# contracts/openapi/scene-api-v1.yaml
openapi: 3.0.3
...
info:
  title: scene-api
  version: 1.2.0
  description: |
    SceneTrip 앱 **작품검색 탭**과 **경로여정 탭**을 받치는 API. 작품·촬영지·인물을
    검색하고, 지도 뷰포트 안의 촬영지를 조회하고, 방문할 장소를 장바구니에 담고,
    담은 장소로 며칠짜리 코스를 짜서 남과 나눈다.

    ## 언어

    사람이 읽는 모든 문자열(작품 제목, 장소명, 주소, 장면 설명)은
    `Accept-Language` 헤더를 따른다. 해당 언어의 번역이 없으면 `ko` 로 폴백하고,
    실제로 무엇을 골랐는지 `Content-Language` 응답 헤더로 알린다.

    ## 검색 범위

    **통합검색이다.** 사용자는 검색 전에 탭을 고르지 않는다. 검색어에 걸린 것과
    이어진 것을 **작품과 장소 양쪽으로 모두** 돌려주고, 탭은 그것을 종류별로 나눠
    보여 줄 뿐이다.

    훑는 곳은 여섯이고, 두 엔드포인트가 **같은 텍스트 뭉치**를 본다.

    | 훑는 곳 | 어떻게 |
    | --- | --- |
    | 장소명 · 장소 별칭 | `search_term` |
    | 장소 설명 | 원문 부분 일치 |
    | 작품 제목 · 작품 별칭 | `search_term` |
    | 작품 설명 | 원문 부분 일치 |
    | 인물 이름 (배우·감독) | `search_term` |
    | **장면 설명** (작품 × 장소) | 원문 부분 일치 |

    **장면 설명은 작품 하나와 장소 하나의 조합에 붙는다.** "2화에서 지은탁이 등장하는
    버스정류장" 처럼 *어느 작품의 어느 장면이 여기서 찍혔는가* 를 적은 문장이라,
    작품에도 장소에도 속하지 않고 그 둘을 잇는 자리에 있다. 그래서 여기에 걸리면
    **작품 탭과 장소 탭이 함께 채워진다.**

    걸린 것은 `place_content` 와 `content_cast` 를 타고 **양쪽으로 옮겨진다.**

    | 검색어 | 작품 탭 | 장소 탭 |
    | --- | --- | --- |
    | `도깨비` (작품명) | 도깨비 | 도깨비 촬영지 전부 |
    | `북촌한옥마을` (장소명) | 거기서 찍은 작품 전부 | 북촌한옥마을 |
    | `공유` (인물명) | 공유 출연작 | 그 출연작의 촬영지 전부 |
    | `한강` (장소 설명) | 그 장소의 촬영작 | 설명에 그 말이 든 장소 |
    | `버스정류장` (장면 설명) | 그 장면이 나온 작품 | 그 장면을 찍은 장소 |

    별칭이 있으므로 `Goblin` 으로도 `도깨비` 가 걸린다.

    **다리는 하나만 건넌다.** `도깨비` 로 그 촬영지들까지는 오지만, 그 촬영지에서
    찍은 *또 다른* 작품까지 딸려 오지는 않는다. 무한히 번지면 결과가 사실상 전체
    목록이 되어 검색이 뜻을 잃는다.

...

...

...

    ## 인증

...

...
servers:
...
paths:
...
  /courses:
    get:
      tags: [courses]
      operationId: listCourses
      summary: 내 코스 목록
      description: |
        여행 중(`active`)인 코스가 위, 예정(`upcoming`)이 아래이고 각각 최근에 만든 순이다.
        홈 화면의 정렬 그대로라 프론트는 받은 순서를 그리면 된다.
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
        - $ref: "#/components/parameters/InstallId"
      responses:
        "200":
          description: 코스 목록. 하나도 없으면 빈 배열이다.
          headers:
            Content-Language:
              $ref: "#/components/headers/ContentLanguage"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [courses]
      operationId: createCourse
      summary: 코스 만들기
      description: |
        **받는 것은 기간과 만든 방식뿐이다.** 떠나는 날도 시작 시각도 받지 않는다.

        만들어진 코스는 일차만 있고 장소는 비어 있다. 장소를 담는 것은 편집이고,
        편집은 `PUT /courses/{courseId}` 한 번으로 끝난다.

        마켓에서 담아 오는 경우는 이 경로가 아니라
        `POST /market/courses/{marketCourseId}/saves` 다.
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
        - $ref: "#/components/parameters/InstallId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseCreate"
      responses:
        "201":
          description: 만들어진 코스. 일차는 비어 있다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /courses/{courseId}:
    get:
      tags: [courses]
      operationId: getCourse
      summary: 코스 상세
      description: |
        일차와 그 안의 장소를 순서대로, 하루 소요 시간까지 담아 돌려준다. 화면 하나를
        그리는 데 필요한 것이 전부 이 응답에 있다.

        **편집 화면도 이것 하나로 시작한다.** 받은 내용을 프론트가 메모리에 들고
        편집하다가, 완료를 누를 때 `PUT` 으로 되돌려 보낸다.
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
        - $ref: "#/components/parameters/InstallId"
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: 코스 상세
          headers:
            Content-Language:
              $ref: "#/components/headers/ContentLanguage"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      tags: [courses]
      operationId: replaceCourse
      summary: 편집 완료 — 코스 전체 교체
      description: |
        **편집 화면의 「완료」가 부르는 하나뿐인 요청이다.**

        제목을 고치고, 기간을 늘렸다 줄이고, 장소를 담고 빼고, 드래그로 순서를 바꾸고,
        체류시간을 조절하는 동안 서버로는 아무것도 나가지 않는다. 프론트가 코스 사본을
        들고 편집하다가 완료를 누를 때 **최종 모습을 통째로** 보낸다.

        그래서 되돌리기가 공짜다 — 잘못 지웠어도 완료 전이면 서버에 아무 일도 일어나지
        않았다. 그리고 한 트랜잭션이라 반쪽만 반영되는 상태가 없다.

        ## 보낸 것이 전부다

        | 보낸 것 | 결과 |
        | --- | --- |
        | `days` 배열의 길이 | 그것이 곧 기간이다. 3개면 3일 코스 |
        | `id` 가 있는 아이템 | 원래 있던 것 — **그대로 유지된다** |
        | `id` 가 없는 아이템 | 새로 담은 것 |
        | 요청에 없는 아이템 | **지워진다** (직접 찍은 핀이었다면 핀도 함께) |
        | `startDate` 를 뺌 | 떠나는 날을 지운다 |

        **기존 장소는 받았던 `id` 를 반드시 실어 보낸다.** 빠뜨리면 서버가 새 장소로
        보아 **방문 체크(`visitedAt`)가 날아간다.**

        ## 기간을 줄이면 장소가 지워진다

        5일 코스의 `days` 를 3개만 보내면 4·5일차의 장소가 **삭제된다.** 되돌릴 수 없다.

        **프론트는 완료를 누르기 전에 확인을 받아야 한다** — "4일차·5일차의 장소 3곳이
        삭제됩니다". 편집 화면이 이미 코스 전체를 들고 있으므로 몇 곳인지 세는 데 추가
        호출이 필요 없다.

        ## 상태는 여기서 못 바꾼다

        코스를 시작하거나 일차를 넘기는 것은 편집이 아니라 여행 중 동작이다.
        `PUT /courses/{courseId}/progress` 를 쓴다.
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
        - $ref: "#/components/parameters/InstallId"
        - $ref: "#/components/parameters/CourseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseReplace"
      responses:
        "200":
          description: 반영된 코스. 이동시간과 하루 소요 시간이 새 순서로 다시 계산돼 있다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: |
            여행 중인 코스를 지금 일차보다 짧게 줄이려 했다. 3일차를 걷고 있는데 2일
            코스로 만들 수는 없다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [courses]
      operationId: deleteCourse
      summary: 코스 삭제
      description: |
        일차·장소·이 코스에서 찍은 핀이 함께 지워진다.

        **마켓에 올려 둔 사본은 남는다.** 사본이라 원본과 수명이 다르다 — 내리려면
        `DELETE /market/courses/{marketCourseId}` 를 따로 부른다.
      parameters:
        - $ref: "#/components/parameters/InstallId"
        - $ref: "#/components/parameters/CourseId"
      responses:
        "204":
          description: 삭제됨
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /courses/{courseId}/progress:
    put:
      tags: [courses]
      operationId: updateCourseProgress
      summary: 코스 시작·일차 이동 (여행 중)
      description: |
        **편집이 아니라 여행 중 동작이라 즉시 반영된다.** 「시작」을 누르면 `active` 가
        되고 그때부터 방문 체크와 지도 위 현재 위치가 나온다.

        상태에 「완료」가 없는 것은 목업 그대로다. 여행을 끝내면 `upcoming` 으로 돌아간다.

        `active` 로 바꿀 때 `currentDayNo` 를 함께 보낸다. `upcoming` 으로 되돌릴 때는
        보내지 않는다 — 예정 코스에 "지금 몇 일차" 는 뜻이 없다.
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
        - $ref: "#/components/parameters/InstallId"
        - $ref: "#/components/parameters/CourseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseProgress"
      responses:
        "200":
          description: 바뀐 코스
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /courses/{courseId}/items/{itemId}/visit:
    put:
      tags: [courses]
      operationId: updateCourseItemVisit
      summary: 방문 체크 (여행 중)
      description: |
        성지에 도착해 체크를 누르는 동작이다. **편집이 아니라 즉시 반영된다** — 이때
        코스 전체를 보내라고 할 수는 없다.

        토글이라 이미 체크된 것을 다시 체크해도 오류가 아니다. 체크한 시각은 서버가
        찍는다.
      parameters:
        - $ref: "#/components/parameters/InstallId"
        - $ref: "#/components/parameters/CourseId"
        - $ref: "#/components/parameters/ItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitUpdate"
      responses:
        "204":
          description: 반영됨
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: 여행 중이 아닌 코스. 예정 상태에서는 방문 체크가 뜻이 없다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: 응답 언어. 생략하거나 해당 번역이 없으면 `ko` 로 폴백한다.
      schema:
        $ref: "#/components/schemas/Lang"
    InstallId:
      name: X-Install-Id
      in: header
      required: true
      description: |
        앱이 최초 실행 시 만들어 보관하는 **설치 UUID**. 기기가 아니라 그 설치본을
        가리킨다 — 앱을 지웠다 깔면 새 값이 된다.

        저장의 주체는 이 값이 아니라 이 값으로 찾아낸 계정이다. 서버가 처음 보는
        값이면 비회원 계정을 만들어 짝지어 둔다.
      schema:
        type: string
        format: uuid
...
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    ItemId:
      name: itemId
      in: path
      required: true
      description: 코스 안의 장소 하나. `CourseItem.id` 다.
      schema:
        type: integer
        format: int64
...
  headers:
    ContentLanguage:
      description: 서버가 실제로 고른 언어. 요청한 언어와 다를 수 있다(폴백).
      schema:
        $ref: "#/components/schemas/Lang"
  responses:
    BadRequest:
      description: 파라미터가 규칙에 맞지 않음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: 대상이 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    InternalError:
      description: |
        서버가 요청을 처리하지 못함. **재시도해도 된다** — 클라이언트가 고칠 것은 없다.

        이 응답에만 `traceId` 가 실린다. 사용자가 그 문자열 하나를 알려 주면 서버 쪽에서
        그 요청의 트레이스와 로그를 그대로 찾을 수 있으므로, **화면에 보여 주거나 오류
        보고에 함께 담는다.** 다른 상태 코드에는 이 필드가 없다.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    Lang:
      type: string
      enum: [ko, en, ja, zh-Hant]
      default: ko
      description: 지원 언어. `zh-Hant` 는 번체(대만·홍콩)로 간체와 구분한다.
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            기계가 분기하는 식별자. 클라이언트 로직은 이 값만 보고 판단한다.

            **전체 목록과 각 코드의 뜻은 `docs/api/errors.md` 에 있다.** 여기에
            같은 목록을 두면 한쪽만 고쳐져 어긋나므로 옮겨 적지 않는다.
          example: PLACE_NOT_FOUND
        message:
          type: string
          description: 사람이 읽는 설명. 사용자에게 그대로 보여 줄 문구가 아니다.
        traceId:
          type: string
          nullable: true
          description: |
            이 요청의 추적 ID. SigNoz 에서 같은 값으로 조회할 수 있다.

            **`500`·`503` 응답에만 실린다.** `400`·`404`·`409`·`422` 는 클라이언트가 무엇을 잘못했는지
            `code` 로 이미 알 수 있어 서버 로그를 뒤질 일이 없다. 값이 없는 필드는 응답에서
            아예 빠지므로, 다른 상태 코드에서는 이 키가 보이지 않는다.
...
    CourseStatus:
      type: string
      enum: [upcoming, active]
      description: |
        예정(`upcoming`) / 여행 중(`active`).

        **이 값이 여행 중 기능의 스위치다.** `active` 일 때만 방문 체크와 길찾기가
        열린다. 「완료」 상태가 없는 것은 목업 그대로다 — 여행을 끝내면 `upcoming` 으로
        돌아가고, 지난 여행은 떠나는 날로 구분한다.
    CourseOrigin:
      type: string
      enum: [ai, self, market]
      description: |
        이 코스를 어떻게 만들었는가. AI 가 짠 코스와 직접 짠 코스의 완주율이 다를 것이라
        나중에 비교하려면 만든 시점에 남겨 둬야 한다 — 소급이 안 된다.

        `market` 은 코스마켓에서 담아 온 것이고, 이 값은 서버가 붙인다.
    CoursePace:
      type: string
      enum: [tight, loose]
      description: |
        「빡빡하게 / 널널하게」. AI 로 짠 코스만 값이 있다.

        **지금은 두 값의 결과가 같다.** 항목은 확정됐지만 일정을 어떻게 바꾸는지가 아직
        정해지지 않았다(3주차 회의 Open Issue 3). 그래도 받아 두는 이유는 로직이
        정해지면 지난 코스도 다시 계산할 수 있어서다.
    PinCategory:
      type: string
      enum: [lodging, food, attraction, street, building]
      description: |
        직접 찍은 핀의 분류. 목업의 다섯 갈래다 — 숙소 · 음식점·카페 · 명소·자연 ·
        거리·다리 · 건물·시설.

        등록된 촬영지의 `type` 과 다르다. 그쪽은 수집된 자유 문자열이고 이쪽은 사용자가
        고르는 닫힌 목록이다.
    TravelBasis:
      type: string
      enum: [straight-line, route]
      description: |
        이동시간을 무엇으로 냈는가. `straight-line` 은 직선거리에서 나온 **추정**이라
        실제와 다르다. 계획 단계는 언제나 이 값이다.
    CourseItemSource:
      type: string
      enum: [place, customPin]
      description: 등록된 촬영지인가, 사용자가 지도에 직접 찍은 핀인가.
    CourseCreate:
      type: object
      required: [dayCount, origin]
      properties:
        dayCount:
          type: integer
          minimum: 1
          maximum: 15
          description: 며칠짜리인가. 이만큼의 빈 일차가 함께 만들어진다.
        title:
          type: string
          maxLength: 50
          description: 비우면 서버가 「이름 없는 코스」로 채운다. 응답에는 언제나 값이 있다.
        origin:
          $ref: "#/components/schemas/CourseOrigin"
        pace:
          $ref: "#/components/schemas/CoursePace"
    CourseList:
      type: object
      required: [items, totalCount]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CourseSummary"
        totalCount:
          type: integer
    CourseSummary:
      type: object
      required: [id, title, dayCount, status, origin, placeCount, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          description: 비어 있지 않다. 사용자가 안 적었으면 서버가 채운 값이다.
        dayCount:
          type: integer
          description: 며칠짜리인가. 편집에서 보낸 `days` 배열의 길이가 그대로 여기 온다.
        startDate:
          type: string
          format: date
          nullable: true
          description: |
            떠나는 날. 정하지 않았으면 없다. 돌아오는 날은 주지 않는다 —
            `startDate + dayCount - 1` 로 나오는 파생값이다.
        status:
          $ref: "#/components/schemas/CourseStatus"
        currentDayNo:
          type: integer
          nullable: true
          description: 여행 중일 때 지금 몇 일차인지. 예정 코스에는 없다.
        origin:
          $ref: "#/components/schemas/CourseOrigin"
        pace:
          $ref: "#/components/schemas/CoursePace"
        sourceMarketCourseId:
          type: integer
          format: int64
          nullable: true
          description: 마켓에서 담아 왔다면 그 원본. 「담아 온 코스」 표시에 쓴다.
        placeCount:
          type: integer
          description: 모든 일차의 장소 수 합계. 목록 카드의 "장소 7곳" 이다.
        thumbnailUrl:
          type: string
          nullable: true
          format: uri
          description: 첫 번째 장소의 대표 이미지. 장소가 하나도 없으면 비어 있다.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CourseDetail:
      allOf:
        - $ref: "#/components/schemas/CourseSummary"
        - type: object
          required: [days]
          properties:
            days:
              type: array
              description: |
                1일차부터 순서대로. **빈 일차도 빠지지 않는다** — 배열 길이가 곧
                기간이다.
              items:
                $ref: "#/components/schemas/CourseDay"
    CourseDay:
      type: object
      required: [dayNumber, items, dwellMinutes, travelMinutes, totalMinutes, travelBasis]
      properties:
        dayNumber:
          type: integer
          minimum: 1
          description: |
            1 부터. **날짜가 아니다** — 떠나는 날이 없어도 일차는 있다. 배열 순서와 같다.
        items:
          type: array
          description: 갈 순서대로. 프론트는 받은 순서를 그대로 그리면 된다.
          items:
            $ref: "#/components/schemas/CourseItem"
        dwellMinutes:
          type: integer
          description: 이 날 머무는 시간의 합
        travelMinutes:
          type: integer
          description: |
            이 날 이동에 드는 시간의 합. **일차 합계로만 준다** — 구간별 소요 시간은
            여행 전에 보여 주지 않기로 했으므로 `CourseItem` 에 그 필드가 아예 없다.
        totalMinutes:
          type: integer
          description: |
            `dwellMinutes + travelMinutes`. **시작 시각을 받지 않으므로 도착·출발 시각은
            없다.** 화면은 "약 6시간 20분" 처럼 길이만 보여 준다.
        travelBasis:
          $ref: "#/components/schemas/TravelBasis"
    CourseItem:
      type: object
      required: [id, source, name, latitude, longitude, dwellMinutes]
      properties:
        id:
          type: integer
          format: int64
          description: |
            **편집 완료(`PUT`)에서 이 값을 그대로 돌려보내야 한다.** 빠뜨리면 서버가
            새 장소로 보아 방문 체크가 날아간다.
        source:
          $ref: "#/components/schemas/CourseItemSource"
        placeId:
          type: integer
          format: int64
          nullable: true
          description: |
            등록된 촬영지일 때만 있다. 직접 찍은 핀에는 없다 — 그 핀은 `place` 에 들어가지
            않아 촬영지 상세로 갈 수 없다.
        customPinId:
          type: integer
          format: int64
          nullable: true
          description: |
            직접 찍은 핀일 때만 있다. **편집에서 이 값을 쓸 일은 없다** — 자리를 가리키는
            것은 아이템의 `id` 이고, 핀은 그 자리 전용이라 함께 따라간다. 같은 숙소를 두
            자리에 넣었다면 이 값이 서로 다르다.
        name:
          type: string
        address:
          type: string
          nullable: true
          description: 등록된 촬영지만 있다. 직접 찍은 핀은 주소를 받지 않는다.
        category:
          type: string
          nullable: true
          description: |
            등록된 촬영지는 수집된 유형 문자열, 직접 찍은 핀은 `PinCategory` 값이다.
            둘의 값 범위가 다르므로 하나로 묶어 필터에 쓰지 않는다.
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        imageUrl:
          type: string
          nullable: true
          format: uri
        dwellMinutes:
          type: integer
          minimum: 15
          maximum: 180
          description: 머무는 시간(분). 15분 ~ 3시간이 목업 슬라이더의 범위다.
        distanceMetersFromPrevious:
          type: integer
          nullable: true
          description: |
            앞 장소에서 여기까지의 **직선거리**다. 첫 번째 장소에는 없다.

            실제 도로 거리가 아니고, 여기에 대응하는 소요 시간 필드도 없다 — 여행 전에는
            "몇 분" 을 보여 주지 않는다. 직접 찍은 핀도 이 계산에 들어간다(숙소가 하루의
            시작과 끝이라 빼면 합계가 크게 어긋난다).
        sourceContentId:
          type: integer
          format: int64
          nullable: true
          description: |
            어떤 작품 때문에 담았는지. 장바구니에서 넘어올 때 따라온다.

            같은 N서울타워가 이태원 클라쓰의 것인지 눈물의 여왕의 것인지 구별하는 값이라,
            화면에 작품 배지를 띄울 때 이것을 쓴다.
        sourceContentTitle:
          type: string
          nullable: true
        visitedAt:
          type: string
          format: date-time
          nullable: true
          description: |
            여행 중 방문 체크를 누른 시각. 없으면 아직 안 갔다는 뜻이다.

            **편집 완료에서 `id` 를 빠뜨리면 이 값이 사라진다.**
    CourseReplace:
      type: object
      required: [title, days]
      description: |
        편집 화면의 최종 모습. **보낸 것이 코스의 전부이고, 빠진 것은 지운 것이다.**
      properties:
        title:
          type: string
          maxLength: 50
          description: 비우면 서버가 「이름 없는 코스」로 채운다.
        startDate:
          type: string
          format: date
          description: 떠나는 날. **빼면 지운다.**
        days:
          type: array
          minItems: 1
          maxItems: 15
          description: |
            1일차부터 순서대로. **배열 길이가 곧 기간이다** — 5일 코스에 3개를 보내면
            3일 코스가 되고 4·5일차의 장소는 삭제된다.

            빈 일차는 `items` 가 빈 배열인 항목으로 표현한다. 일차는 큐라서 뒤에서만
            늘고 준다 — 가운데를 빼는 조작은 없다.
          items:
            $ref: "#/components/schemas/CourseDayInput"
    CourseDayInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          description: |
            갈 순서대로. **배열 순서가 곧 방문 순서다** — 드래그 결과를 따로 보내는
            요청이 없는 이유가 이것이다.
          items:
            $ref: "#/components/schemas/CourseItemInput"
    CourseItemInput:
      type: object
      description: |
        `placeId` 와 `customPin` 중 **정확히 하나**를 보낸다. 둘 다이거나 둘 다 없으면
        `400` 이다. OpenAPI 로는 이 조건을 표현할 방법이 마땅치 않아 서버가 검사한다.
      properties:
        id:
          type: integer
          format: int64
          description: |
            **원래 있던 장소면 조회에서 받은 `id` 를 그대로 넣는다.** 그래야 방문
            체크(`visitedAt`)가 보존된다. 새로 담은 장소는 이 값을 비운다.
        placeId:
          type: integer
          format: int64
          description: |
            등록된 촬영지. 장바구니에서 꺼내든 검색으로 찾든 같은 값이다.
        customPin:
          $ref: "#/components/schemas/CustomPinInput"
        dwellMinutes:
          type: integer
          minimum: 15
          maximum: 180
          description: |
            머무는 시간(분).

            **비우면 서버가 장소 유형에 맞는 기본값을 채운다** — 고궁 90, 카페 40,
            다리 15 처럼 다르고 모르는 유형은 60 이다. 프론트가 그 표를 들고 있을 필요가
            없다. 값이 필요하면 담은 뒤 응답의 `CourseItem.dwellMinutes` 를 보면 된다.

            사용자가 슬라이더를 만졌으면 그 값을 보낸다. 이미 담긴 장소를 옮기기만 할
            때도 지금 값을 그대로 실어 보내는 편이 안전하다 — 비우면 기본값으로
            되돌아간다.
        sourceContentId:
          type: integer
          format: int64
          description: |
            어떤 작품 때문에 담았는지. 장바구니에서 넘어올 때 그 값을 그대로 옮긴다.
    CustomPinInput:
      type: object
      required: [name, category, latitude, longitude]
      description: |
        등록되지 않은 곳을 지도에 직접 찍은 핀. 주로 숙소다. 이름·분류·위경도만 받는다 —
        주소도 작품도 없다.

        **핀은 담은 자리마다 따로 만들어진다.** 2박 3일 숙소를 1일차 끝과 2일차 끝에
        넣으면 핀이 둘 생긴다. 하나를 여러 자리가 나눠 쓰게 하면 프론트가 핀에 이름표를
        붙여 관리해야 하는데, 그 복잡함에 비해 얻는 것이 없다고 봤다 — 사용자 눈에는
        어느 쪽이든 같은 화면이다.

        그 대신 **숙소 이름이나 위치를 고치면 담은 자리마다 각각 보내야 한다.** 프론트가
        한 곳에서 고쳤다면 그 값을 해당 자리 전부에 채워 보낸다.

        `id` 가 있는 아이템에 이 값을 함께 보내면 그 자리의 핀을 갱신한다. 아이템이
        요청에서 빠지면 그 핀도 함께 지워진다 — 핀이 그 자리 전용이라 남길 이유가 없다.

        **이 핀은 그 코스 안에서만 존재한다.** 검색·지도 결과에 나타나지 않고 다른
        코스로 가져갈 수도 없다. 마켓에 올릴 때도 빠진다.
      properties:
        name:
          type: string
          maxLength: 50
        category:
          $ref: "#/components/schemas/PinCategory"
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
    CourseProgress:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/CourseStatus"
        currentDayNo:
          type: integer
          minimum: 1
          description: |
            지금 몇 일차인지. `active` 로 바꿀 때 함께 보낸다. `upcoming` 으로 되돌릴
            때는 보내지 않는다 — 예정 코스에는 뜻이 없는 값이다.
    VisitUpdate:
      type: object
      required: [visited]
      properties:
        visited:
          type: boolean
          description: 체크했으면 `true`, 되돌리면 `false`. 시각은 서버가 찍는다.

    # ───────────── 작품 찜 ─────────────
...